requirements.txt
certifi==2018.4.16
emoji==0.5.0
future==0.16.0
python-telegram-bot==10.0.1
Wand==0.4.4
